requirements.txt
numpy
Pillow
PyOpenGL
PyQt5
